(#13) Refactor UI model into focused modules
