+++ internal/ui/render_helpers.go
@@ -1,0 +1,249 @@
+package ui
+
+import (
+	"fmt"
+	"sort"
+	"strings"
+
+	"github.com/adrunkhuman/90minuTUI/internal/site"
+)
+
+func renderSeasonsWindow(seasons []site.Season, cursor int) []string {
+	if len(seasons) == 0 {
+		return []string{"(none)"}
+	}
+
+	start, end := windowBounds(len(seasons), cursor, 10)
+	lines := make([]string, 0, end-start)
+	for i := start; i < end; i++ {
+		prefix := "  "
+		if i == cursor {
+			prefix = "> "
+		}
+
+		marker := ""
+		if seasons[i].Current {
+			marker = " *"
+		}
+
+		lines = append(lines, fmt.Sprintf("%s%s%s", prefix, seasons[i].Label, marker))
+	}
+
+	return lines
+}
+
+func renderCompetitionWindow(items []site.Competition, cursor int) []string {
+	if len(items) == 0 {
+		return []string{"(none)"}
+	}
+
+	start, end := windowBounds(len(items), cursor, 18)
+	lines := make([]string, 0, end-start)
+	for i := start; i < end; i++ {
+		prefix := "  "
+		if i == cursor {
+			prefix = "> "
+		}
+		lines = append(lines, prefix+items[i].Name)
+	}
+
+	return lines
+}
+
+func windowBounds(total, cursor, maxItems int) (int, int) {
+	if total <= maxItems {
+		return 0, total
+	}
+
+	half := maxItems / 2
+	start := cursor - half
+	if start < 0 {
+		start = 0
+	}
+
+	end := start + maxItems
+	if end > total {
+		end = total
+		start = end - maxItems
+	}
+
+	return start, end
+}
+
+func clamp(v, minV, maxV int) int {
+	if maxV < minV {
+		return minV
+	}
+	if v < minV {
+		return minV
+	}
+	if v > maxV {
+		return maxV
+	}
+	return v
+}
+
+func renderPlayerLine(player site.PlayerLine) string {
+	line := player.Name
+	if len(player.Events) == 0 {
+		return line
+	}
+
+	line += " [" + strings.Join(player.Events, ", ") + "]"
+	return line
+}
+
+func sortedEvents(events []site.MatchEvent) []site.MatchEvent {
+	ordered := make([]site.MatchEvent, len(events))
+	copy(ordered, events)
+
+	sort.SliceStable(ordered, func(i, j int) bool {
+		mi, hi := minuteSortKey(ordered[i].MinuteText)
+		mj, hj := minuteSortKey(ordered[j].MinuteText)
+
+		if hi != hj {
+			return hi
+		}
+		if hi && mj != mi {
+			return mi < mj
+		}
+
+		weightI := eventWeight(ordered[i].Kind)
+		weightJ := eventWeight(ordered[j].Kind)
+		if weightI != weightJ {
+			return weightI < weightJ
+		}
+
+		return false
+	})
+
+	return ordered
+}
+
+func eventWeight(kind string) int {
+	switch kind {
+	case "GOAL":
+		return 0
+	case "RC":
+		return 1
+	case "YC":
+		return 2
+	case "SUB":
+		return 3
+	default:
+		return 9
+	}
+}
+
+func minuteSortKey(text string) (int, bool) {
+	if text == "" {
+		return 0, false
+	}
+
+	parts := strings.SplitN(text, "+", 2)
+	base := atoiOrNeg(parts[0])
+	if base < 0 {
+		return 0, false
+	}
+
+	extra := 0
+	if len(parts) == 2 {
+		extra = max(0, atoiOrNeg(parts[1]))
+	}
+
+	return base*100 + extra, true
+}
+
+func atoiOrNeg(s string) int {
+	value := 0
+	if _, err := fmt.Sscanf(strings.TrimSpace(s), "%d", &value); err != nil {
+		return -1
+	}
+	return value
+}
+
+func formatEventLabel(event site.MatchEvent) string {
+	prefix := event.Kind
+	if prefix == "" {
+		prefix = "EVENT"
+	}
+
+	if event.Text == "" {
+		return prefix
+	}
+
+	return prefix + " " + event.Text
+}
+
+func renderSideBySide(left, middle, right string, width int) string {
+	if width < 30 {
+		if middle == "" {
+			return left + " | " + right
+		}
+		return left + " | " + middle + " | " + right
+	}
+
+	midWidth := 8
+	gap := 2
+	sideWidth := max(8, (width-midWidth-(gap*2))/2)
+
+	leftText := padRight(truncate(left, sideWidth), sideWidth)
+	midText := padRight(truncate(middle, midWidth), midWidth)
+	rightText := truncate(right, sideWidth)
+
+	return leftText + strings.Repeat(" ", gap) + midText + strings.Repeat(" ", gap) + rightText
+}
+
+func truncate(value string, maxLen int) string {
+	if maxLen <= 0 {
+		return ""
+	}
+	r := []rune(value)
+	if len(r) <= maxLen {
+		return value
+	}
+	if maxLen == 1 {
+		return "…"
+	}
+
+	return string(r[:maxLen-1]) + "…"
+}
+
+func padRight(value string, width int) string {
+	pad := width - len([]rune(value))
+	if pad <= 0 {
+		return value
+	}
+
+	return value + strings.Repeat(" ", pad)
+}
+
+func layoutWidths(total int, collapsed, emphasizeRight bool) (int, int) {
+	if total < 40 {
+		return 0, total
+	}
+
+	if collapsed {
+		return 0, total
+	}
+
+	leftWidth := 36
+	if emphasizeRight {
+		leftWidth = clamp(total/4, 28, 42)
+	} else {
+		leftWidth = clamp(total/3, 32, 50)
+	}
+
+	rightWidth := total - leftWidth - 1
+	if rightWidth < 40 {
+		rightWidth = 40
+		leftWidth = max(0, total-rightWidth-1)
+	}
+
+	if leftWidth < 24 {
+		leftWidth = 0
+		rightWidth = total
+	}
+
+	return leftWidth, rightWidth
+}
